requirements.txt
discord.py==1.7.3
Flask==3.0.3
Requests==2.32.2
